// web/src/api/models/CoolifyWidgetApplication.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mindscape
 * This is the swagger page for the Project mindscape generated with Egg-go. use this to test your database connection
 *
 * The version of the OpenAPI document: 0.0.1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface CoolifyWidgetApplication
 */
export interface CoolifyWidgetApplication {
	/**
	 *
	 * @type {string}
	 * @memberof CoolifyWidgetApplication
	 */
	createdAt?: string;
	/**
	 *
	 * @type {string}
	 * @memberof CoolifyWidgetApplication
	 */
	deletedAt?: string;
	/**
	 *
	 * @type {string}
	 * @memberof CoolifyWidgetApplication
	 */
	description?: string;
	/**
	 *
	 * @type {string}
	 * @memberof CoolifyWidgetApplication
	 */
	fqdn?: string;
	/**
	 *
	 * @type {number}
	 * @memberof CoolifyWidgetApplication
	 */
	id?: number;
	/**
	 *
	 * @type {string}
	 * @memberof CoolifyWidgetApplication
	 */
	name?: string;
	/**
	 *
	 * @type {string}
	 * @memberof CoolifyWidgetApplication
	 */
	redirect?: string;
	/**
	 *
	 * @type {string}
	 * @memberof CoolifyWidgetApplication
	 */
	status?: string;
	/**
	 *
	 * @type {string}
	 * @memberof CoolifyWidgetApplication
	 */
	updatedAt?: string;
	/**
	 *
	 * @type {string}
	 * @memberof CoolifyWidgetApplication
	 */
	uuid?: string;
}

/**
 * Check if a given object implements the CoolifyWidgetApplication interface.
 */
export function instanceOfCoolifyWidgetApplication(
	value: object,
): value is CoolifyWidgetApplication {
	return true;
}

export function CoolifyWidgetApplicationFromJSON(
	json: any,
): CoolifyWidgetApplication {
	return CoolifyWidgetApplicationFromJSONTyped(json, false);
}

export function CoolifyWidgetApplicationFromJSONTyped(
	json: any,
	ignoreDiscriminator: boolean,
): CoolifyWidgetApplication {
	if (json == null) {
		return json;
	}
	return {
		createdAt: json['created_at'] == null ? undefined : json['created_at'],
		deletedAt: json['deleted_at'] == null ? undefined : json['deleted_at'],
		description: json['description'] == null ? undefined : json['description'],
		fqdn: json['fqdn'] == null ? undefined : json['fqdn'],
		id: json['id'] == null ? undefined : json['id'],
		name: json['name'] == null ? undefined : json['name'],
		redirect: json['redirect'] == null ? undefined : json['redirect'],
		status: json['status'] == null ? undefined : json['status'],
		updatedAt: json['updated_at'] == null ? undefined : json['updated_at'],
		uuid: json['uuid'] == null ? undefined : json['uuid'],
	};
}

export function CoolifyWidgetApplicationToJSON(
	json: any,
): CoolifyWidgetApplication {
	return CoolifyWidgetApplicationToJSONTyped(json, false);
}

export function CoolifyWidgetApplicationToJSONTyped(
	value?: CoolifyWidgetApplication | null,
	ignoreDiscriminator: boolean = false,
): any {
	if (value == null) {
		return value;
	}

	return {
		created_at: value['createdAt'],
		deleted_at: value['deletedAt'],
		description: value['description'],
		fqdn: value['fqdn'],
		id: value['id'],
		name: value['name'],
		redirect: value['redirect'],
		status: value['status'],
		updated_at: value['updatedAt'],
		uuid: value['uuid'],
	};
}