// web/src/api/models/CoolifyWidgetService.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mindscape
 * This is the swagger page for the Project mindscape generated with Egg-go. use this to test your database connection
 *
 * The version of the OpenAPI document: 0.0.1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface CoolifyWidgetService
 */
export interface CoolifyWidgetService {
	/**
	 *
	 * @type {string}
	 * @memberof CoolifyWidgetService
	 */
	createdAt?: string;
	/**
	 *
	 * @type {string}
	 * @memberof CoolifyWidgetService
	 */
	deletedAt?: string;
	/**
	 *
	 * @type {string}
	 * @memberof CoolifyWidgetService
	 */
	description?: string;
	/**
	 *
	 * @type {number}
	 * @memberof CoolifyWidgetService
	 */
	id?: number;
	/**
	 *
	 * @type {string}
	 * @memberof CoolifyWidgetService
	 */
	name?: string;
	/**
	 *
	 * @type {string}
	 * @memberof CoolifyWidgetService
	 */
	serviceType?: string;
	/**
	 *
	 * @type {string}
	 * @memberof CoolifyWidgetService
	 */
	updatedAt?: string;
	/**
	 *
	 * @type {string}
	 * @memberof CoolifyWidgetService
	 */
	uuid?: string;
}

/**
 * Check if a given object implements the CoolifyWidgetService interface.
 */
export function instanceOfCoolifyWidgetService(
	value: object,
): value is CoolifyWidgetService {
	return true;
}

export function CoolifyWidgetServiceFromJSON(json: any): CoolifyWidgetService {
	return CoolifyWidgetServiceFromJSONTyped(json, false);
}

export function CoolifyWidgetServiceFromJSONTyped(
	json: any,
	ignoreDiscriminator: boolean,
): CoolifyWidgetService {
	if (json == null) {
		return json;
	}
	return {
		createdAt: json['created_at'] == null ? undefined : json['created_at'],
		deletedAt: json['deleted_at'] == null ? undefined : json['deleted_at'],
		description: json['description'] == null ? undefined : json['description'],
		id: json['id'] == null ? undefined : json['id'],
		name: json['name'] == null ? undefined : json['name'],
		serviceType:
			json['service_type'] == null ? undefined : json['service_type'],
		updatedAt: json['updated_at'] == null ? undefined : json['updated_at'],
		uuid: json['uuid'] == null ? undefined : json['uuid'],
	};
}

export function CoolifyWidgetServiceToJSON(json: any): CoolifyWidgetService {
	return CoolifyWidgetServiceToJSONTyped(json, false);
}

export function CoolifyWidgetServiceToJSONTyped(
	value?: CoolifyWidgetService | null,
	ignoreDiscriminator: boolean = false,
): any {
	if (value == null) {
		return value;
	}

	return {
		created_at: value['createdAt'],
		deleted_at: value['deletedAt'],
		description: value['description'],
		id: value['id'],
		name: value['name'],
		service_type: value['serviceType'],
		updated_at: value['updatedAt'],
		uuid: value['uuid'],
	};
}